Consolidate pool creation by Jupiter and Vintage
This addresses code review feedback. Specifically, it extracts a factory
for creating `ForkJoinPool` instances. It also consolidates the
documentation on customising pool creation. Lastly, the unit tests are
refactored to simplify future enhancements.

Issue: #2229
