+++ documentation/src/docs/asciidoc/user-guide/advanced-topics.adoc
@@ -8,3 +8,43 @@ include::testkit.adoc[]
 ////
 include::engines.adoc[]
 ////
+
+[[advanced-topics-parallel-execution-config]]
+=== Parallel Execution Configuration
+
+When enabling parallel test execution using either the Jupiter or Vintage engines,
+properties such as the desired parallelism and the maximum pool size can be configured
+using a `{ParallelExecutionConfigurationStrategy}`. The JUnit Platform provides two
+implementations out of the box: `dynamic` and `fixed`. Alternatively, you may implement a
+`custom` strategy.
+
+To select a strategy, set the `junit.<engine>.execution.parallel.config.strategy`
+configuration parameter to one of the following options. Replace "engine" with either
+`jupiter` or `vintage`.
+
+`dynamic`::
+  Computes the desired parallelism based on the number of available processors/cores
+  multiplied by the `junit.<engine>.execution.parallel.config.dynamic.factor`
+  configuration parameter (defaults to `1`).
+
+`fixed`::
+  Uses the mandatory `junit.<engine>.execution.parallel.config.fixed.parallelism`
+  configuration parameter as the desired parallelism.
+
+`custom`::
+  Allows you to specify a custom `{ParallelExecutionConfigurationStrategy}`
+  implementation via the mandatory `junit.<engine>.execution.parallel.config.custom.class`
+  configuration parameter to determine the desired configuration.
+
+If no configuration strategy is set, JUnit uses the `dynamic` configuration
+strategy with a factor of `1`. Consequently, the desired parallelism will be equal to the
+number of available processors/cores.
+
+.Parallelism does not imply maximum number of concurrent threads
+NOTE: JUnit Jupiter does not guarantee that the number of concurrently executing tests
+will not exceed the configured parallelism. For example, when using one of the
+synchronization mechanisms described in the
+<<writing-tests-parallel-execution-synchronization,Jupiter Synchronization>>
+section, the `ForkJoinPool` that is used behind the scenes may spawn additional threads to
+ensure execution continues with sufficient parallelism. Thus, if you require such
+guarantees in a test class, please use your own means of controlling concurrency.
